Add an environment variable to give the developer an option to disable apt-get usage
After the introduction of the "set -e" option, the build script fails on non debian/ubuntu linux distributions.

This is a quick fix to allow the users of distros other than debian/ubuntu to build the library, not a definitive solution.
The COCOS2DX_USEAPT environment variable can be set in the user shell config or used directly in
the command line: COCOS2DX_USEAPT=false ./make-all-linux-project.sh