Support "image layer" for Tiled Map (CCTMXTiledMap).
New tests in TileMapTest: TMXOrthoBackgroundOnlyTest, TMXOrthoBackgroundWithSpritesTest.

New classes for CCTMXTiledMap: CCTMXImageLayer, CCTMXImageLayer, CCTMXImageLayerInfo.

New method for CCTMXTiledMap: imageLayerNamed() - Return the TMXImageLayer for the specific layer.

Correct error when tiled map (CCTMXTiledMap) not consists of tilesets.

Correct potential error for parse of the image source of tilesets (added enum TMXPropertyTileSet).

Assert for CCTMXLayer, when texture for tileset is not loaded.