Update CCSpriteFrameCache.cpp
If a PVR texture is used in the texture atlas created by Zwoptex, the sprite frame will not load the PVR, but instead it will try to load the texture as PNG because the "textureFileName" key doesn't contains the file name. It seems that Zwoptex is saving texture name and extension in the "target" key.