Fixes full paths on Android (cleaner)
If one adds a search resolutions order directory without a trailing slash, CCFileUtils::getFullPathForDirectoryAndFilename will fail on Android. For example it will return:

assets/sdBg1.png

Instead of the expected:

assets/sd/Bg1.png

This small commit fixes the function to add a slash if necessary, fixing the issue on Android.

Compared to the last pull request of this name, formatting matches the style of code around it and comments have been added.