[nacl] Ensure uniform precision is the same for shaders within the same program.
Recent version of chrome will fail to link shader programs
that contain the same type names with different precisions.

Previously CCGLProgram::compileShader was delaring the
CC_PMatrix and friends to be highp in the vertex shader
and mediump in the fragment shader. This is technically
against the GLES spec but I assume that most implementation
allow it anyway.

This changes the precision to highp for both vertext
and fragment shaders when building for NaCl.  It might
be desirable to make this true for all platform but I'm
not sure what the consequences would be.

Also, print linker logs when shader programs fail to link.
The fact that this was not previously done made this problem
hard to debug.

Also, the code for checking the link status was guarded by
DEBUG (which AFIAKT is never defined) rather than
COCOS2D_DEBUG.