fix cracked on android when schedule mixed with "TimerTargetCallback" and other "Timer"
when a node scheduled mix with "TimerTargetCallback" that has parameter "key" and other "Timer" that without "key" , it will crack on same android devices at line "if (key == timer->getKey())". 
static_cast is an unsafe conversion from a pointer of a base class to a pointer of a derived class.