Sets up cocos2d-x for 64-bit builds.
By far the biggest change here is in the ccUTF8 files: I reverted them back to
the version we originally checked out when we first forked the repo. The
reason for this is because in the most recent update to ccUTF8, there
seemed to be some changes made for llvm 3.x. I'm not 100% certain of the
reprecussions on there, but it seems to work for us in building 32-bit
and 64-bit versions. I can't really tell if there's anything majorly
wrong here, although we should diligently test (in several languages).

The rest of the changes are simple one-line bug fixes (changing the name
of EAGLView to cc2dEAGLView) and changing various project properties to
build for architectures armv7 and arm64 only.