Fix Sprite3DCache crash/cleanup
`Sprite3DCache` is not being cleaned up properly, so it is possible to create
an instance of it and if you then destroy the Director and try to create
a new one, the new one will use the same `Sprite3DCache` which has
dangling references to the old Director, causing a crash as soon as the
application starts (with the second Director) in this case.

This fixes it by destroying the Sprite3DCache when the director is
purged.