Patch to make edit boxes for iOS behave as expected
Currently there is an issue with edit boxes not scaling correctly for the
interactive OS part. The font is scaled incorrectly and the box itself is
incorrectly sized.

Fixed this by scaling the system control and system control font iOS side.
Added a new function to avoid messing with previous cocos code, as recommended.

Tested on iOS iPhone 6+, Mini iPad retina and Nexus 5.

Buddied with Aaron and Stephen