enhance egl-selectconfig
I notice that when we are using "glContextAttrs = {8, 8, 8, 8, 24, 8}" in AppDelegate, and when using configchooser to choose the config, it will search whether it has the config "8,8,8,8,24,8", if it's not exists then it will select the first config available. But in ASUS Nexus 7, one of the many configs available is "8,8,8,8,16,8", but instead the old codes chooses "8,8,8,8,0,0", and it makes UIScrollView implementation in Nexus 7 become all green and white..

This changes need to be addressed further by testing it onto many devices.