Ensure CCLabel atlases are purged before resetting
The current purge->reset behavior occurs per label according to scene
graph priority.  This has two disadvanages; firstly, any labels that are
initialized but not part of the scene graph will not be reset.
Secondly, any labels which share atlases in the FontAtlasCache may not
be reset since each label will call releaseFontAtlas individually before
resetting its fontConfig.  This means that any shared atlases may not be
released all the way to a reference count of 0 and the invalid atlas
will not be deleted and reset from scratch.

By putting a fixed priority on both the purge and reset steps we can
ensure that firstly: all labels will be reset regardless of whether they
have been added to the scene-graph.  Secondly, ALL labels with the same
atlas must call releaseFontAtlas before any others are allowed to
rebuild their textures.  This will prevent any old atlases from sticking
around after a new GL context is created.

This is a fix for Github issue #8305
https://github.com/cocos2d/cocos2d-x/issues/8305