Rewrote all action ease classes to remove the redundancy of having a single class definition for every possible ease function. This has been done through some template classes which accept functions as arguments.
This makes both header and source file much shorter and much more maintanable. Readability shouldn't be a concern since now there is almost no code to read.

Performance should identical since everything is bound at compile time. The following patch should be compatible with all recent (C++11) compilers but I have no way to test it, I tested it on LLVM3.4 on a OS X machine.

This patch also gives the possiblity to have ready to go custom ease function without the need to sublass ActionEase at all (as long as the prototype of the function is supported).