Fix ClippingRectangleNode scissor positioning
Currently, onBeforeVisitScissor function scales BOTH the position and
size by the nodes world scale.  The position, however, is already in
world coordinates since it is passed through convertToWorldSpace. This
change ensures that a rectangle in the local space of the node which
is passed into setClippingRegion is correctly converted to GL coords
before it is passed to glScissor.