Multi-window implementation
WindowKey is used as a multiplatform identifier for windows.  For now it's a hash of the GLFWwindow pointer, but that can be changed.

DirectorContext encapsulates a "mini director" instance for each window.  There is still a central director which is responsible for updating each context.  Each context also has its own event manager, action manager, and scheduler.

To reduce the amount of code changes necessary for mobile, by default the first window created is the main window.  The main window is special; if it gets closed, it terminates the application.  There are special accessors to grab the "main"scheduler/eventmanager/actionmanager.  For certain tasks, especially async tasks, they need to schedule their callbacks on the main window, because the existence of the other windows is not guaranteed, unlike the main one.

The contexts work in a similar manner as opengl.  A context is set to be active (setContextAndLock) and then non-actve (unlock).  Doing so changes what gets returned by the getCurrentWindow call and the getCurrentScheduler/EventManager/ActionManager calls.  This allows nodes and virtually everything to not care which window it was created in.  Unless the node is transplanted into a new window/context, its getCurrentManagers functions will always return the same.  It also hides any window dependancy.  Almost everything outside of the director doens't know or care which window it was created in.

CCGLViewImpl-desktop.cpp required substantial changes.  The poll events function grabs all the events for all windows at once.  What we want to do is sort them by window and defer the execution of them until the correct context is active.  Otherwise, we'll be switching contexts rapidly while processing the polled events.  Thus, the callbacks all push std::functions onto a map, keyed to the correct window.

There are also a number of order of operation fixes and smaller miscelaneous changes that were required as a result of this feature.

cpp-empty-test has been extended to create 3 windows at launch to demonstrate the multi window functionality.