Fixed windows compile issues.
I'm uncertain what the desired functionality is for many of the functions that try to get the scheduler/eventdispatcher/actionmanager.  In my multi window implementation, you can either get the "current" one (which is the one that's active - something that's typically only true when you're in the body of the Director's draw function), OR you can get the "main" one (which is always associated with the main window and generally guaranteed to exist).  The main ones are required for async calls since the current windows may be different or not exist when the async callbacks get triggered.  The main ones should also be used (stylistically) for "global" systems.