Refactored Android Downloader
Turned DataTaskHandler and FileTaskHandler into inner classes of
Cocos2dxDownloader. These classes are only supposed to be used by
Cococ2dxDownloader instances and this way they do not need to be passed
a reference to said instances explicitly.

Removed DownloadTask as it seems to be unnecessary.

Deprecated the use of static methods to create instances and doing work
on instances.