Lua: fix cc.AABB:updateMinMax
This function was unusable because it had been written to expect an
object based on a C++ cocos2d::Vec3, whereas elsewhere cc.vec3 objects
are held on the lua side as simple tables. We have changed this so now
it can be called with a lua array of tables as one would expect. Since
the lua arrray carries it's length, the separate count argument is not
now needed