Remove unnecessary arguments from some lua cc.mat4 methods (#16658)
* Lua: remove unused 3rd argument from cc.mat4.transformVector

cc.mat4.transformVector takes a mat4 and a 4vector and returns the result
of applying the mat4 to the vector. Strangely it had been made to also
require a third argument called dst which it doesn't use but did check
was a vec4. This commit allows cc.mat4.transformVector to be called
without the unnecessary argument. We still permit calling with it for
backward compatibility.

Also changed is the wrapper in Cocos2d.lua so that a second form of
cc.mat4.transformVector can be used where the vector is supplied as
4 separate numbers.

Changed the test files to call without the unnecessary argument.

* Lua: remove unused argument from cc.mat4.createTranslation/Rotation

Both these functions have an extra unnecessary argument called dst,
although neither use it either to pass in or out a value. This
commit removes the extra argument and updates the test programs
correspondingly.

Also, within Cocos2d.lua, remove some definitions that are later
overwritten by new ones