EventListenerTouchOneByOne::_claimedTouches cleanup bugfix
Bug description:
Touch pointer won't be added to `_claimedTouches` in `EventListenerTouchOneByOne` if touch listener became unregistered after `onTouchBegan` call, but it still wll be added if it was disabled or paused. And so it won't be removed in future.
For example if I will remove node connected with listener from scene onTouchBegan, but won't destroy it and add to scene after some time, than It's listener will keep invalid pointer to old touch (because listener won't be processed by dispatcher on ending or cancelling touch).