Use CMake 3.14's first-class support for iOS cross-compilation.
CMake 3.14 has built-in support for creating iOS/tvOS/watchOS Xcode projects without the need for a custom toolchain script.

See: https://cmake.org/cmake/help/v3.14/manual/cmake-toolchains.7.html#cross-compiling-for-ios-tvos-or-watchos

To build for an iOS device:

	~ mkdir ios-build
	~ cd ios-build
	~ cmake .. -GXcode -DCMAKE_SYSTEM_NAME=iOS -DCMAKE_OSX_SYSROOT=iphoneos

To build for the iOS Simulator:

	~ mkdir ios-build
	~ cd ios-build
	~ cmake .. -GXcode -DCMAKE_SYSTEM_NAME=iOS -DCMAKE_OSX_SYSROOT=iphonesimulator

To build for macOS:

	~ mkdir mac-build
	~ cd mac-build
	~ cmake .. -GXcode -DCMAKE_SYSTEM_NAME=Darwin

To run Metal code on the iOS Simulator you must set the deployment target to iOS 13.

Note: At the time of writing the cpp-tests project is building for iOS Simulator however it is crashing due to a shader compile error.
It runs without issue on an iOS 13 device.