Fix error with update line height (#19914)
After movig font atlas `reinit()` (https://github.com/cocos2d/cocos2d-x/pull/19384) there was indused a problem: the first display of text with a outline in the label does not update line height. This is because the atlas `reinit()` does not occur when the outline is activated. In this case, a new atlas will be created, but the line height will not be taken from it, because line height calculate in `updateContent()`. I propose line height calculating to return to the font atlas constructor.