Fix incorrect rendering with tilemaps with csv data on windows
Each tilemap row would be shifted to the right a certain number of
tiles.
This is caused by the carriage return in windows being '\r\n' and
the way each csv line is parsed.
The '\r' at the beginning of each line would be parsed as a tile id.
On windows we must remove all 'r' from the csv string.