wordSeparatorMode
Same languages like Spanish has space (32) as word separator.  Other like Chinese has no separator.

The current implementation based the decision of which char is a word separator on the location of the device. Some games let the user decide the language from a settings, so it is good to let them th choice.

CJK unicodes has a len of 3 or more.

I added these option for the developer to choose from:

WordSeparatorBasedOnDevice:   current implementation

WordSeparatorSpaceSlashNotHighUnicode:  Spaces and '-' will be consider words separators.  Recommended for Korean

WordSeparatorSpaceSlashAndKCJ: Spaces, '-' and KCJ unicodes will be consider words separators.  Recommended for Spanish

You can see the test on:
Node:UI > GUI Dynamic Create Test > RichTest Test > language

The fnction findSplitPositionForWord receives the parameter wordSeparatorMode and chooses the inlinen function (charWrappableFunctionPointer) to be use in getNextWordPos and then in std::find_if to look for a good spot to but the sentence.